fix: resolve all Android compilation errors across 15 files
- Downgrade Firebase BOM from 34.15.0 to 33.7.0 (missing analytics-ktx artifact)
- Add missing Theme aliases: Red, Danger colors + H3 text style
- Fix DeviceStateClassifier: return DeviceState enum, add battery/alarm intents
- Fix MqttManager: MqttClient -> MqttAsyncClient with async connect wait
- Fix SplashScreen alpha receiver chain
- Add missing imports: tween, Text, verticalScroll, collectAsStateWithLifecycle
- Fix type mismatches: DeviceScreenState->DeviceUiState, Long->Int stats
- Fix unresolved references: Workspace->Work icon, Slate800/Slate700 colors
- Fix AppNavHost EntryPointAccessors import path
- Fix LogsScreen: use AppColors.Error/Success/Warning
- Remove obsolete ExperimentalFoundationApi annotation
- Fix DeviceRepository classify() -> .name for String API field
- Add isBatteryOptimized to DeviceUiState

Also update backend MQTT ACL config.

+++ android/app/src/main/java/com/aeroxebee/client/ui/screens/dashboard/DashboardScreen.kt
@@ -8,6 +8,7 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.CheckCircle
 import androidx.compose.material.icons.filled.Error
@@ -22,6 +23,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import androidx.compose.material3.Text
 import com.aeroxebee.client.ui.components.*
 import com.aeroxebee.client.ui.theme.*
 
@@ -58,8 +60,8 @@ fun DashboardScreen(
             Column {
                 // ─── Connection Status Hero ─────────────────
                 ConnectionStatusCard(
-                    sentCount = state.stats.sent,
-                    totalMessages = state.stats.total,
+                    sentCount = state.stats.sent.toInt(),
+                    totalMessages = state.stats.total.toInt(),
                 )
 
                 Spacer(Modifier.height(AppSpacing.XXXL))
@@ -107,9 +109,9 @@ fun DashboardScreen(
 
                 // ─── Total Messages ─────────────────────────
                 TotalMessagesCard(
-                    total = state.stats.total,
-                    sent = state.stats.sent,
-                    failed = state.stats.failed,
+                    total = state.stats.total.toInt(),
+                    sent = state.stats.sent.toInt(),
+                    failed = state.stats.failed.toInt(),
                 )
             }
         }
